change Milliseconds return type

+++ src/core/lib/event_engine/time_util.h
@@ -17,6 +17,7 @@
 #include <grpc/support/port_platform.h>
 
 #include <stdint.h>
+#include <vcruntime.h>
 
 #include <grpc/event_engine/event_engine.h>
 
@@ -24,7 +25,7 @@ namespace grpc_event_engine {
 namespace experimental {
 
 // Convert a duration to milliseconds
-int64_t Milliseconds(EventEngine::Duration d);
+size_t Milliseconds(EventEngine::Duration d);
 
 }  // namespace experimental
 }  // namespace grpc_event_engine
